Move config to static; move manager packages to controller
