+++ internal/framework/controller/index/endpointslice.go
