+++ internal/framework/controller/predicate/service.go
