+++ internal/mode/static/config/config.go
