+++ internal/mode/static/manager.go
@@ -16,13 +16,13 @@ import (
 	k8spredicate "sigs.k8s.io/controller-runtime/pkg/predicate"
 	gatewayv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 
-	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/config"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/controller"
+	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/controller/filter"
+	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/controller/index"
+	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/controller/predicate"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/events"
-	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/manager/filter"
-	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/manager/index"
-	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/manager/predicate"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/status"
+	"github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/config"
 	ngxcfg "github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/nginx/config"
 	ngxvalidation "github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/nginx/config/validation"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/nginx/file"
