+++ internal/mode/static/state/change_processor_test.go
@@ -17,8 +17,8 @@ import (
 	"sigs.k8s.io/gateway-api/apis/v1beta1"
 
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/conditions"
+	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/controller/index"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/helpers"
-	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/manager/index"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/state"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/state/graph"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/state/relationship"
