+++ internal/mode/static/state/relationship/capturer.go
@@ -9,7 +9,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/gateway-api/apis/v1beta1"
 
-	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/manager/index"
+	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/controller/index"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/state/graph"
 )
