+++ internal/mode/static/state/relationship/capturer_test.go
@@ -10,8 +10,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/gateway-api/apis/v1beta1"
 
+	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/controller/index"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/helpers"
-	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/manager/index"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/state/relationship"
 )
