+++ internal/mode/static/state/resolver/service_resolver_test.go
@@ -13,8 +13,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
+	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/controller/index"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/helpers"
-	"github.com/nginxinc/nginx-kubernetes-gateway/internal/framework/manager/index"
 	"github.com/nginxinc/nginx-kubernetes-gateway/internal/mode/static/state/resolver"
 )
